Prevent error when qmlimportscanner not available
When using `make qmake_all` the qmlimportscanner is not
built when creating Makefiles for targets depending on
QML so the error must be ignored.

BTW: Using `make qmake_all` is required to fix Makefiles
using sed because qmake fails to set library search paths
for host standard library correctly.

Not sure why this error does not occur when cross
compiling static Qt for Windows.