Allow keeping Android app in background with QtQuickView
Do not terminate Qt as this would stop the main Qt thread which is not
desirable when the app is supposed to keep running in the background (and
still supposed to process events via the Qt event loop, e.g. to process
network traffic and update a notification accordingly). This change also
makes sure that the absence of an activity or service is treated as if the
app is not a "full" Qt application as we're still just using QtQuickView
here.

Change-Id: I18e2c2a054b19bdbdbe03d189f9a36dcdbe1e204